[CHANGE] Configuration Checkstyle + BD de test, correction des erreurs checkfile, Debut des tests unitaires (DAO)

+++ src/main/java/com/excilys/cdb/dao/DAO.java
@@ -4,57 +4,68 @@ import java.sql.SQLException;
 
 import com.excilys.cdb.model.Computer;
 import com.excilys.cdb.utils.Page;
-import java.sql.SQLException;
+
 /**
- * Interface regroupant toutes les transactions sur la BD
- * 
- * @param <T> L'objet manipulé
+ * Interface regroupant toutes les transactions sur la BD.
+ * @param <T>
+ *            L'objet manipulé
  */
 public interface DAO<T> {
-	
-	/**
-	 * Récupère une page d'objet de la base
-	 * @param page le numéro de la page
-	 * @return la page contenant les objets
-	 * @throws SQLException Exception liée à la requete
-	 */
-	  Page<T> findAll(int page) throws SQLException;
-	  
-	  /**
-	   * Récupère l'objet voulu
-	   * @param id l'id de l'objet voulu
-	   * @return l'objet voulu
-	   * @throws SQLException Exception liée à la requete
-	   */
-	  T findById(int id) throws SQLException;
-	  
-	  /**
-	   * Ajoute une objet dans la base
-	   * @param t l'objet à ajouter
-	 * @return 
-	   * @throws SQLException Exception liée à la requete
-	   */
-	  int add(T t) throws SQLException;
-	  
-	  /**
-	   * Supprime l'objet voulu
-	   * @param id l'id de l'objet à supprimer
-	   * @throws SQLException Exception liée à la requete
-	   */
-	  void delete(int id) throws SQLException;
-	  
-	  /**
-	   * Met à jour l'objet voulu
-	   * @param t l'objet contenant les nouvelles informations
-	 * @return 
-	   * @throws SQLException Exception liée à la requete
-	   */
-	  Computer update(T t) throws SQLException;
-	  
-	  /**
-	   * Récupère le nombre de page maximum d'un table d'objet
-	   * @return le nombre maximum de page
-	   * @throws SQLException Exception liée à la requete
-	   */
-	  int getMaxPage() throws SQLException;
+
+    /**
+     * Récupère une page d'objet de la base.
+     * @param page
+     *            le numéro de la page
+     * @return la page contenant les objets
+     * @throws SQLException
+     *             Exception liée à la requete
+     */
+    Page<T> findAll(int page) throws SQLException;
+
+    /**
+     * Récupère l'objet voulu.
+     * @param id
+     *            l'id de l'objet voulu
+     * @return l'objet voulu
+     * @throws SQLException
+     *             Exception liée à la requete
+     */
+    T findById(long id) throws SQLException;
+
+    /**
+     * Ajoute une objet dans la base.
+     * @param t
+     *            l'objet à ajouter
+     * @return l'id du nouvelle objet
+     * @throws SQLException
+     *             Exception liée à la requete
+     */
+    int add(T t) throws SQLException;
+
+    /**
+     * Supprime l'objet voulu.
+     * @param id
+     *            l'id de l'objet à supprimer
+     * @throws SQLException
+     *             Exception liée à la requete
+     */
+    void delete(long id) throws SQLException;
+
+    /**
+     * Met à jour l'objet voulu.
+     * @param t
+     *            l'objet contenant les nouvelles informations
+     * @return le nouvel objet
+     * @throws SQLException
+     *             Exception liée à la requete
+     */
+    Computer update(T t) throws SQLException;
+
+    /**
+     * Récupère le nombre de page maximum d'un table d'objet.
+     * @return le nombre maximum de page
+     * @throws SQLException
+     *             Exception liée à la requete
+     */
+    int getMaxPage() throws SQLException;
 }
